Move examplebroker.go to the examplebroker package
Now examplebroker is an internal package with all of its config files
and the dbus mapping
