+++ internal/brokers/examplebroker/examplebroker.go
@@ -1,4 +1,5 @@
-package brokers
+// Package examplebroker implements an example broker that will be used by the authentication daemon.
+package examplebroker
 
 import (
 	"context"
@@ -13,6 +14,7 @@ import (
 	"time"
 
 	"github.com/google/uuid"
+	"github.com/ubuntu/authd/internal/brokers"
 	"golang.org/x/exp/slices"
 )
 
@@ -28,7 +30,7 @@ type isAuthorizedCtx struct {
 	cancelFunc context.CancelFunc
 }
 
-type exampleBroker struct {
+type broker struct {
 	currentSessions        map[string]sessionInfo
 	currentSessionsMu      sync.RWMutex
 	userLastSelectedMode   map[string]string
@@ -74,8 +76,8 @@ const (
 	brokerEncryptionKey = "encryptionkey"
 )
 
-func newExampleBroker(name string) (b *exampleBroker, fullName, brandIcon string) {
-	return &exampleBroker{
+func newBroker(name string) (b *broker, fullName, brandIcon string) {
+	return &broker{
 		currentSessions:        make(map[string]sessionInfo),
 		currentSessionsMu:      sync.RWMutex{},
 		userLastSelectedMode:   make(map[string]string),
@@ -86,7 +88,7 @@ func newExampleBroker(name string) (b *exampleBroker, fullName, brandIcon string
 }
 
 // NewSession creates a new session for the specified user.
-func (b *exampleBroker) NewSession(ctx context.Context, username, lang string) (sessionID, encryptionKey string, err error) {
+func (b *broker) NewSession(ctx context.Context, username, lang string) (sessionID, encryptionKey string, err error) {
 	sessionID = uuid.New().String()
 	b.currentSessionsMu.Lock()
 	b.currentSessions[sessionID] = sessionInfo{
@@ -98,7 +100,7 @@ func (b *exampleBroker) NewSession(ctx context.Context, username, lang string) (
 }
 
 // GetAuthenticationModes returns the list of supported authentication modes for the selected broker depending on session info.
-func (b *exampleBroker) GetAuthenticationModes(ctx context.Context, sessionID string, supportedUILayouts []map[string]string) (authenticationModes []map[string]string, err error) {
+func (b *broker) GetAuthenticationModes(ctx context.Context, sessionID string, supportedUILayouts []map[string]string) (authenticationModes []map[string]string, err error) {
 	sessionInfo, err := b.sessionInfo(sessionID)
 	if err != nil {
 		return nil, err
@@ -258,7 +260,8 @@ func (b *exampleBroker) GetAuthenticationModes(ctx context.Context, sessionID st
 	return authenticationModes, nil
 }
 
-func (b *exampleBroker) SelectAuthenticationMode(ctx context.Context, sessionID, authenticationModeName string) (uiLayoutInfo map[string]string, err error) {
+// SelectAuthenticationMode returns the UI layout information for the selected authentication mode.
+func (b *broker) SelectAuthenticationMode(ctx context.Context, sessionID, authenticationModeName string) (uiLayoutInfo map[string]string, err error) {
 	// Ensure session ID is an active one.
 	sessionInfo, err := b.sessionInfo(sessionID)
 	if err != nil {
@@ -303,7 +306,7 @@ func (b *exampleBroker) SelectAuthenticationMode(ctx context.Context, sessionID,
 }
 
 // IsAuthorized evaluates the provided authenticationData and returns the authorisation level of the user.
-func (b *exampleBroker) IsAuthorized(ctx context.Context, sessionID, authenticationData string) (access, infoUser string, err error) {
+func (b *broker) IsAuthorized(ctx context.Context, sessionID, authenticationData string) (access, infoUser string, err error) {
 	sessionInfo, err := b.sessionInfo(sessionID)
 	if err != nil {
 		return "", "", err
@@ -344,73 +347,73 @@ func (b *exampleBroker) IsAuthorized(ctx context.Context, sessionID, authenticat
 }
 
 //nolint:unparam // This is an static example implementation, so we don't return an error other than nil.
-func (b *exampleBroker) handleIsAuthorized(ctx context.Context, sessionInfo sessionInfo, authData map[string]string) (access, infoUser string, err error) {
+func (b *broker) handleIsAuthorized(ctx context.Context, sessionInfo sessionInfo, authData map[string]string) (access, infoUser string, err error) {
 	// Note that the "wait" authentication can be cancelled and switch to another mode with a challenge.
 	// Take into account the cancellation.
 	switch sessionInfo.selectedMode {
 	case "password":
 		if authData["challenge"] != "goodpass" {
-			return AuthDenied, "", nil
+			return brokers.AuthDenied, "", nil
 		}
 
 	case "pincode":
 		if authData["challenge"] != "4242" {
-			return AuthDenied, "", nil
+			return brokers.AuthDenied, "", nil
 		}
 
 	case "totp_with_button", "totp":
 		if authData["challenge"] != "temporary pass" {
-			return AuthDenied, "", nil
+			return brokers.AuthDenied, "", nil
 		}
 
 	case "phoneack1":
 		if authData["wait"] != "true" {
-			return AuthDenied, "", nil
+			return brokers.AuthDenied, "", nil
 		}
 		// Send notification to phone1 and wait on server signal to return if OK or not
 		select {
 		case <-time.After(5 * time.Second):
-			return AuthDenied, "", nil
+			return brokers.AuthDenied, "", nil
 		case <-ctx.Done():
-			return AuthCancelled, "", nil
+			return brokers.AuthCancelled, "", nil
 		}
 
 	case "phoneack2":
 		if authData["wait"] != "true" {
-			return AuthDenied, "", nil
+			return brokers.AuthDenied, "", nil
 		}
 
 		// This one is failing remotely as an example
 		select {
 		case <-time.After(2 * time.Second):
-			return AuthDenied, "", nil
+			return brokers.AuthDenied, "", nil
 		case <-ctx.Done():
-			return AuthCancelled, "", nil
+			return brokers.AuthCancelled, "", nil
 		}
 
 	case "fidodevice1":
 		if authData["wait"] != "true" {
-			return AuthDenied, "", nil
+			return brokers.AuthDenied, "", nil
 		}
 
 		// simulate direct exchange with the FIDO device
 		select {
 		case <-time.After(5 * time.Second):
-			return AuthDenied, "", nil
+			return brokers.AuthDenied, "", nil
 		case <-ctx.Done():
-			return AuthCancelled, "", nil
+			return brokers.AuthCancelled, "", nil
 		}
 
 	case "qrcodewithtypo":
 		if authData["wait"] != "true" {
-			return AuthDenied, "", nil
+			return brokers.AuthDenied, "", nil
 		}
 		// Simulate connexion with remote server to check that the correct code was entered
 		select {
 		case <-time.After(4 * time.Second):
-			return AuthDenied, "", nil
+			return brokers.AuthDenied, "", nil
 		case <-ctx.Done():
-			return AuthCancelled, "", nil
+			return brokers.AuthCancelled, "", nil
 		}
 	}
 
@@ -420,32 +423,32 @@ func (b *exampleBroker) handleIsAuthorized(ctx context.Context, sessionInfo sess
 		if authData["challenge"] != "" {
 			// validate challenge given manually by the user
 			if authData["challenge"] != "aaaaa" {
-				return AuthDenied, "", nil
+				return brokers.AuthDenied, "", nil
 			}
 		} else if authData["wait"] == "true" {
 			// we are simulating clicking on the url signal received by the broker
 			// this can be cancelled to resend a challenge
 			select {
 			case <-time.After(10 * time.Second):
-				return AuthDenied, "", nil
+				return brokers.AuthDenied, "", nil
 			case <-ctx.Done():
-				return AuthCancelled, "", nil
+				return brokers.AuthCancelled, "", nil
 			}
 		} else {
-			return AuthDenied, "", nil
+			return brokers.AuthDenied, "", nil
 		}
 	}
 
 	infoUser, exists := users[sessionInfo.username]
 	if !exists {
-		return AuthDenied, "", nil
+		return brokers.AuthDenied, "", nil
 	}
 
-	return AuthAllowed, infoUser, nil
+	return brokers.AuthAllowed, infoUser, nil
 }
 
 // EndSession ends the requested session and triggers the necessary clean up steps, if any.
-func (b *exampleBroker) EndSession(ctx context.Context, sessionID string) error {
+func (b *broker) EndSession(ctx context.Context, sessionID string) error {
 	if _, err := b.sessionInfo(sessionID); err != nil {
 		return err
 	}
@@ -463,7 +466,7 @@ func (b *exampleBroker) EndSession(ctx context.Context, sessionID string) error 
 
 // CancelIsAuthorized cancels the IsAuthorized request for the specified session.
 // If there is no pending IsAuthorized call for the session, this is a no-op.
-func (b *exampleBroker) CancelIsAuthorized(ctx context.Context, sessionID string) {
+func (b *broker) CancelIsAuthorized(ctx context.Context, sessionID string) {
 	b.isAuthorizedCallsMu.Lock()
 	defer b.isAuthorizedCallsMu.Unlock()
 	if _, exists := b.isAuthorizedCalls[sessionID]; !exists {
@@ -534,7 +537,7 @@ func decryptAES(key []byte, ct string) string {
 }
 
 // sessionInfo returns the session information for the specified session ID or an error if the session is not active.
-func (b *exampleBroker) sessionInfo(sessionID string) (sessionInfo, error) {
+func (b *broker) sessionInfo(sessionID string) (sessionInfo, error) {
 	b.currentSessionsMu.RLock()
 	defer b.currentSessionsMu.RUnlock()
 	session, active := b.currentSessions[sessionID]
